move tracefs test
